+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxAuxData.hs
@@ -3,22 +3,35 @@
 
 module Cardano.Ledger.Conway.TxAuxData () where
 
+import Cardano.Ledger.Alonzo.Core
 import Cardano.Ledger.Alonzo.TxAuxData (
-  AlonzoTxAuxData,
+  AlonzoTxAuxData (..),
   hashAlonzoTxAuxData,
+  metadataAlonzoTxAuxDataL,
+  plutusScriptsAllegraTxAuxDataL,
+  timelockScriptsAlonzoTxAuxDataL,
   translateAlonzoTxAuxData,
   validateAlonzoTxAuxData,
  )
 import Cardano.Ledger.Conway.Era
 import Cardano.Ledger.Conway.Scripts ()
-import Cardano.Ledger.Core
 import Cardano.Ledger.Crypto
 
 instance Crypto c => EraTxAuxData (ConwayEra c) where
   type TxAuxData (ConwayEra c) = AlonzoTxAuxData (ConwayEra c)
+
+  mkBasicTxAuxData = AlonzoTxAuxData mempty mempty mempty
+
+  metadataTxAuxDataL = metadataAlonzoTxAuxDataL
 
   upgradeTxAuxData = translateAlonzoTxAuxData
 
   hashTxAuxData = hashAlonzoTxAuxData
 
   validateTxAuxData = validateAlonzoTxAuxData
+
+instance Crypto c => AllegraEraTxAuxData (ConwayEra c) where
+  timelockScriptsTxAuxDataL = timelockScriptsAlonzoTxAuxDataL
+
+instance Crypto c => AlonzoEraTxAuxData (ConwayEra c) where
+  plutusScriptsTxAuxDataL = plutusScriptsAllegraTxAuxDataL