+++ eras/mary/impl/src/Cardano/Ledger/Mary/TxAuxData.hs
@@ -12,7 +12,12 @@ module Cardano.Ledger.Mary.TxAuxData (
 )
 where
 
-import Cardano.Ledger.Allegra.TxAuxData
+import Cardano.Ledger.Allegra.TxAuxData (
+  AllegraEraTxAuxData (..),
+  AllegraTxAuxData (..),
+  metadataAllegraTxAuxDataL,
+  timelockScriptsAllegraTxAuxDataL,
+ )
 import Cardano.Ledger.AuxiliaryData (AuxiliaryDataHash (..))
 import Cardano.Ledger.Core (EraTxAuxData (..), upgradeScript)
 import Cardano.Ledger.Crypto (Crypto)
@@ -27,8 +32,15 @@ import Control.DeepSeq (deepseq)
 instance Crypto c => EraTxAuxData (MaryEra c) where
   type TxAuxData (MaryEra c) = AllegraTxAuxData (MaryEra c)
 
+  mkBasicTxAuxData = AllegraTxAuxData mempty mempty
+
+  metadataTxAuxDataL = metadataAllegraTxAuxDataL
+
   upgradeTxAuxData (AllegraTxAuxData md scripts) = AllegraTxAuxData md $ upgradeScript <$> scripts
 
   validateTxAuxData _ (AllegraTxAuxData md as) = as `deepseq` all validMetadatum md
 
   hashTxAuxData aux = AuxiliaryDataHash (hashAnnotated aux)
+
+instance Crypto c => AllegraEraTxAuxData (MaryEra c) where
+  timelockScriptsTxAuxDataL = timelockScriptsAllegraTxAuxDataL