+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/AuxData.hs
@@ -1,5 +1,7 @@
 module Cardano.Ledger.Api.Tx.AuxData (
   EraTxAuxData (TxAuxData),
+  mkBasicTxAuxData,
+  metadataTxAuxDataL,
   upgradeTxAuxData,
   hashTxAuxData,
   validateTxAuxData,
@@ -9,17 +11,22 @@ module Cardano.Ledger.Api.Tx.AuxData (
   Metadatum (..),
 
   -- * Allegra
+  AllegraEraTxAuxData,
+  timelockScriptsTxAuxDataL,
   AllegraTxAuxData (..),
 
   -- * Alonzo
+  AlonzoEraTxAuxData,
+  plutusScriptsTxAuxDataL,
   AlonzoTxAuxData (..),
   mkAlonzoTxAuxData,
   getAlonzoTxAuxDataScripts,
 )
 where
 
-import Cardano.Ledger.Allegra.TxAuxData (AllegraTxAuxData (..))
+import Cardano.Ledger.Allegra.TxAuxData (AllegraEraTxAuxData (..), AllegraTxAuxData (..))
 import Cardano.Ledger.Alonzo.TxAuxData (
+  AlonzoEraTxAuxData (..),
   AlonzoTxAuxData (..),
   getAlonzoTxAuxDataScripts,
   mkAlonzoTxAuxData,